# backend/service-go/app/service-main/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/apps/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Main API
    version: 0.0.1
paths:
    /service/main/v1/getAvailableNFTList:
        get:
            tags:
                - Main
            operationId: Main_GetAvailableNFTList
            parameters:
                - name: walletPub
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetAvailableNFTListResp'
    /service/main/v1/getLeaderboard:
        get:
            tags:
                - Main
            operationId: Main_GetLeaderboard
            parameters:
                - name: first
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: type
                  in: query
                  schema:
                    type: string
                - name: startTime
                  in: query
                  schema:
                    type: string
                - name: endTime
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetLeaderboardResp'
    /service/main/v1/ping:
        get:
            tags:
                - Main
            operationId: Main_Ping
            parameters:
                - name: msg
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PingResp'
components:
    schemas:
        AvailableNFT:
            type: object
            properties:
                rank_type:
                    type: integer
                    format: int64
                rank_year:
                    type: integer
                    format: int64
                rank_season:
                    type: integer
                    format: int64
                rank:
                    type: integer
                    format: int64
        GetAvailableNFTListResp:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/AvailableNFT'
        GetLeaderboardResp:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/LeaderboardItem'
                type:
                    type: string
        LeaderboardItem:
            type: object
            properties:
                wallet_pub:
                    type: string
                uname:
                    type: string
                retired_amount:
                    type: string
                ens:
                    type: string
                twitter:
                    type: string
                is_contract:
                    type: boolean
        PingResp:
            type: object
            properties:
                res:
                    type: string
tags:
    - name: Main